fix lyx::mkdir 2nd argument.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@434 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/support/lyxlib.h
@@ -36,7 +36,7 @@ namespace lyx {
 	///
 	void abort();
 	///
-	int mkdir(char const * pathname, unsigned int mode);
+	int mkdir(char const * pathname, unsigned long int mode);
 	///
 	int putenv(char const * str);
 }
@@ -57,7 +57,7 @@ struct lyx {
 	///
 	static void abort();
 	///
-	static int mkdir(char const * pathname, unsigned int mode);
+	static int mkdir(char const * pathname, unsigned long int mode);
 	///
 	static int putenv(char const * str);
 };
